Backend/requirements.txt
pyodbc
fastapi
python-dotenv